Move snapshot testsuite to sub package
Prevents imports of snapshot to require importing test packages.
